Feat/chat leave and pw

+++ backend/src/chat/chat.gateway.ts
@@ -35,12 +35,12 @@ export class ChatGateway implements OnModuleInit, OnGatewayDisconnect {
 		});
 	}
 
-	async handleDisconnect(client: Socket) {
-		const userID = this.chatGatewayService.getUserIdFromSocket(client);
-		await this.chatGatewayService.setUserOnlineStatus(userID, false)
-		await this.chatGatewayService.deleteUserSocketID(userID);
-		this.chatGatewayService.deleteUserFromList(userID);
-	}
+  async handleDisconnect(client: Socket) {
+    const userID = this.chatGatewayService.getUserIdFromSocket(client);
+    if (!userID) return;
+	await this.chatGatewayService.setUserOnlineStatus(userID, false)
+    this.chatGatewayService.deleteUserFromList(userID);
+  }
 
 	@SubscribeMessage('connectMessage')
 	async establishConnect(@ConnectedSocket() client: Socket, @MessageBody() data: EstablishConnectDTO) {
